stamp: rename mod
