+++ crates/base/src/lib.rs
@@ -16,4 +16,4 @@ pub use sb_core::cache::CacheSetting;
 pub use sb_graph::DecoratorType;
 
 #[cfg(test)]
-mod tracing;
+mod tracing_subscriber;
